增加nav测试代码
